Programmatically get peer address info.
No longer store addressinfo during connect-socket.
This reduces foreign memory floating around and also makes it easy to get the peer address for accepted connections too.
A side effect is that the connection record is now only a thin wrapper for a socket file descriptor and has been renamed to sockobj.