[syntax] alloc: cleanup and wrap contents in dynamic-wind
Wrap contents in dynamic-wind. It's simpler than what was there previously, handles (values ...) returns, and ensures that memory is free'd even if exceptions are raised.

Remove the lock/unlock object calls. They were added due to a mis-read of the Chez manual as they're only necessary for locking scheme objects whose memory addresses are passed to the C side and therefore need to stay immovable in memory. eg, scheme callback functions.

Define 'alloc' with the one pattern. This removes the duplicated code for count and non-count cases as was there before.